+++ packages/trip-details/src/TripDetails.story.tsx
@@ -13,7 +13,6 @@ import styled from "styled-components";
 import TripDetails, { FareLegTable } from ".";
 import * as TripDetailsClasses from "./styled";
 import {
-  CaloriesDetailsProps,
   DepartureDetailsProps,
   FareDetailsProps,
   FareTableLayout,
@@ -210,17 +209,6 @@ const CustomDepartureDetails = ({
   </>
 );
 
-const CustomCaloriesDetails = ({
-  bikeSeconds,
-  calories,
-  walkSeconds
-}: CaloriesDetailsProps): ReactElement => (
-  <>
-    Custom message about {calories} calories burned,
-    {walkSeconds} seconds and {bikeSeconds} seconds.
-  </>
-);
-
 /**
  * Create a template component for each TripDetails story.
  * A Component argument is passed. Once bound,
@@ -232,7 +220,6 @@ function createTripDetailsTemplate(
 ): ComponentStory<typeof TripDetails> {
   const TripDetailsTemplate = (
     {
-      CaloriesDetails,
       DepartureDetails,
       FareDetails,
       fareDetailsLayout,
@@ -249,7 +236,6 @@ function createTripDetailsTemplate(
       : {};
     return (
       <Component
-        CaloriesDetails={CaloriesDetails}
         DepartureDetails={DepartureDetails}
         FareDetails={FareDetails}
         fareDetailsLayout={fareDetailsLayout}
@@ -377,7 +363,6 @@ export const TncTransitItinerary = makeStory(
 
 export const TncTransitItineraryWithCustomMessages = makeStory(
   {
-    CaloriesDetails: CustomCaloriesDetails,
     defaultFareKey: "electronicRegular",
     DepartureDetails: CustomDepartureDetails,
     itinerary: tncTransitTncItinerary