+++ packages/trip-details/src/types.ts
@@ -3,11 +3,6 @@
 import type { MassUnitOption, Fare, Itinerary, Money } from "@opentripplanner/types";
 import type { ReactElement } from "react";
 
-export interface CaloriesDetailsProps {
-  bikeSeconds: number;
-  calories: number;
-  walkSeconds: number;
-}
 
 export interface CO2ConfigType {
   carbonIntensity?: Record<string, number>;
@@ -63,10 +58,6 @@ export interface TransitFareProps {
 
 export interface TripDetailsProps {
   /**
-   * Slot for a custom component to render the expandable section for calories.
-   */
-  CaloriesDetails?: React.ElementType<CaloriesDetailsProps>;
-  /**
    * Used for additional styling with styled components for example.
    */
   className?: string;
@@ -78,11 +69,6 @@ export interface TripDetailsProps {
    * Slot for a custom component to render the expandable section for departure.
    */
   DepartureDetails?: React.ElementType<DepartureDetailsProps>;
-  /**
-   * If this is set to true, a row will be added to the trip details displaying how
-   * many calories were burned on the active legs of the trip.
-   */
-  displayCalories?: boolean;
   /**
    * Slot for a custom component to render the expandable section for fares.
    */